requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy
pandas==1.3.3
plotly==5.3.1
pynytimes==0.7.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1
requests==2.26.0
six==1.16.0
tenacity==8.0.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.5.0
